+++ lib/iris/tests/unit/cube/test_Cube.py
@@ -1354,123 +1354,6 @@ class Test_dtype(tests.IrisTest):
             self.assertTrue(cube.has_lazy_data())
 
 
-class Test_data_dtype_fillvalue(tests.IrisTest):
-    def setUp(self):
-        self.fill_value = 1234
-
-    def _sample_data(self, dtype=('f4'), masked=False, fill_value=None,
-                     lazy=False):
-        data = np.arange(6).reshape((2, 3))
-        dtype = np.dtype(dtype)
-        data = data.astype(dtype)
-        if masked:
-            data = ma.masked_array(data, mask=[[0, 1, 0], [0, 0, 0]],
-                                   fill_value=fill_value)
-        if lazy:
-            data = as_lazy_data(data)
-        return data
-
-    def _sample_cube(self, dtype=('f4'), masked=False, fill_value=None,
-                     lazy=False):
-        data = self._sample_data(dtype=dtype, masked=masked,
-                                 fill_value=fill_value, lazy=lazy)
-        cube = Cube(data)
-        return cube
-
-    def test_realdata_change(self):
-        # Check re-assigning real data.
-        cube = self._sample_cube()
-        self.assertEqual(cube.dtype, np.float32)
-        new_dtype = np.dtype('i4')
-        new_data = self._sample_data(dtype=new_dtype)
-        cube.data = new_data
-        self.assertIs(cube.core_data(), new_data)
-        self.assertEqual(cube.dtype, new_dtype)
-
-    def test_realmaskdata_change(self):
-        # Check re-assigning real masked data.
-        cube = self._sample_cube(masked=True, fill_value=self.fill_value)
-        self.assertEqual(cube.dtype, np.float32)
-        new_dtype = np.dtype('i4')
-        new_fill_value = 4321
-        new_data = self._sample_data(masked=True,
-                                     fill_value=new_fill_value,
-                                     dtype=new_dtype)
-        cube.data = new_data
-        self.assertIs(cube.core_data(), new_data)
-        self.assertEqual(cube.dtype, new_dtype)
-        self.assertEqual(cube.data.fill_value, new_fill_value)
-
-    def test_lazydata_change(self):
-        # Check re-assigning lazy data.
-        cube = self._sample_cube(lazy=True)
-        self.assertEqual(cube.core_data().dtype, np.float32)
-        new_dtype = np.dtype('f8')
-        new_data = self._sample_data(new_dtype, lazy=True)
-        cube.data = new_data
-        self.assertIs(cube.core_data(), new_data)
-        self.assertEqual(cube.dtype, new_dtype)
-
-    def test_lazymaskdata_change(self):
-        # Check re-assigning lazy masked data.
-        cube = self._sample_cube(masked=True, fill_value=self.fill_value,
-                                 lazy=True)
-        self.assertEqual(cube.core_data().dtype, np.float32)
-        new_dtype = np.dtype('f8')
-        new_fill_value = 4321
-        new_data = self._sample_data(dtype=new_dtype, masked=True,
-                                     fill_value=new_fill_value, lazy=True)
-        cube.data = new_data
-        self.assertIs(cube.core_data(), new_data)
-        self.assertEqual(cube.dtype, new_dtype)
-        self.assertEqual(cube.data.fill_value, new_fill_value)
-
-    def test_lazydata_realise(self):
-        # Check touching lazy data.
-        fill_value = 27.3
-        cube = self._sample_cube(lazy=True)
-        data = cube.data
-        self.assertIs(cube.core_data(), data)
-        self.assertEqual(cube.dtype, np.float32)
-
-    def test_lazymaskdata_realise(self):
-        # Check touching masked lazy data.
-        fill_value = 27.3
-        cube = self._sample_cube(masked=True, fill_value=fill_value, lazy=True)
-        data = cube.data
-        self.assertIs(cube.core_data(), data)
-        self.assertEqual(cube.dtype, np.float32)
-        self.assertArrayAllClose(data.fill_value, fill_value)
-
-    def test_realmaskedconstantint_realise(self):
-        masked_data = ma.masked_array([666], mask=True)
-        masked_constant = masked_data[0]
-        cube = Cube(masked_constant)
-        data = cube.data
-        self.assertTrue(ma.isMaskedArray(data))
-        self.assertNotIsInstance(data, ma.core.MaskedConstant)
-
-    def test_lazymaskedconstantint_realise(self):
-        dtype = np.dtype('i2')
-        masked_data = ma.masked_array([666], mask=True, dtype=dtype)
-        masked_constant = masked_data[0]
-        masked_constant_lazy = as_lazy_data(masked_constant)
-        cube = Cube(masked_constant_lazy)
-        data = cube.data
-        self.assertTrue(ma.isMaskedArray(data))
-        self.assertNotIsInstance(data, ma.core.MaskedConstant)
-
-    def test_lazydata___getitem__dtype(self):
-        dtype = np.dtype('int16')
-        masked_array = ma.masked_array(np.arange(5),
-                                       mask=[0, 0, 1, 0, 0],
-                                       dtype=dtype)
-        lazy_masked_array = as_lazy_data(masked_array)
-        cube = Cube(lazy_masked_array)
-        subcube = cube[3:]
-        self.assertEqual(subcube.dtype, dtype)
-
-
 class TestSubset(tests.IrisTest):
     def test_scalar_coordinate(self):
         cube = Cube(0, long_name='apricot', units='1')
